// src/clients/web3/types.ts
export enum Connector {
  MetaMask = 'MetaMask',
  TrustWallet = 'TrustWallet',
  WalletConnect = 'WalletConnect',
  CoinbaseWallet = 'CoinbaseWallet',
  BinanceChainWallet = 'BinanceChainWallet',
  BraveWallet = 'BraveWallet',
  InfinityWallet = 'InfinityWallet',
  OperaWallet = 'OperaWallet',
  BitKeep = 'BitKeep'
}